Create StatefullSet
What is a Kubernetes StatefulSet?
Kubernetes StatefulSet is an API-object that’s purpose-built to support stateful application components. 
It creates a set of identically configured Pods from a spec you supply, but each Pod is assigned a 
non-interchangeable identity. Pods retain their identity if they have to be rescheduled or you scale the StatefulSet.